requirements.txt
numpy==1.17.4
pandas==0.25.3
mysql-connector-python==8.0.21
python-dotenv==0.14.0
psycopg2==2.8.6
python-dateutil==2.8.1
openpyxl==3.0.5